requirements.txt
playwright==1.44.0
openai==1.57.0
pillow==10.3.0
requests==2.32.3
schedule==1.2.1
python-dotenv==1.0.1
cloudinary==1.40.0
